Added a knockback effect to the punch, messed with physics.
I tried multiple things this morning such as physics. One of which was
to get the punch sprite to stay in front of the player sprite. I failed,
but still have some ideas.

Main change: Added a knock back effect to the punch. It could be
improved, and may be the way we can implement the stomp. Also added a
test object for physics, and can be deleted.