WIP GUI Overhaul and Partial Arena
Made modifications to the way the GUI is displayed; renames the life
controller to a GUI controller, removed cool down objects and started
drawing them through the GUI controller. Added viscera decals as well as
imported vomitoria sprites. At a stable state now, but I'm going to eat,
will be back in the repo later.